arduino : fix bug in the Reed-Solomon code that causes crashes on some MCUs
The issue was first reported here:

https://github.com/ggerganov/ggwave-arduino/issues/1

We were incorrectly reading the "log" array in RS::gf::pow() function.

It is surprising how this code even worked on the other microcontrollers.
Probably we have been reading bogus values, so I expect the performance of
the transmissions to improve after this change.